Rename files MLEstimateComponentBasedNormalisation
