+++ src/recon_buildblock/CMakeLists.txt
@@ -61,7 +61,7 @@ set(${dir_LIB_SOURCES}
 	BinNormalisationFromAttenuationImage.cxx
 	BinNormalisationSPECT.cxx
         BinNormalisationPETFromComponents.cxx
-        ML_estimate_component_based_normalisation.cxx
+        MLEstimateComponentBasedNormalisation.cxx
 	GeneralisedPrior.cxx
 	ProjDataRebinning.cxx
 	FourierRebinning.cxx
